Move FaqsSection to paid ads shared component and reuse in edit ads.
