+++ js/src/setup-ads/ads-stepper/create-campaign/form-content.js
@@ -9,7 +9,7 @@ import { __ } from '@wordpress/i18n';
 import StepContentFooter from '.~/components/stepper/step-content-footer';
 import AudienceSection from '.~/components/paid-ads/audience-section';
 import BudgetSection from './budget-section';
-import FaqsSection from './faqs-section';
+import FaqsSection from '../../../components/paid-ads/faqs-section';
 
 const FormContent = ( props ) => {
 	const { formProps, submitButton } = props;
